добавил stream api при выгрузке данных из БД

+++ src/main/java/com/example/IterationProject/Controllers/MainController.java
@@ -38,7 +38,7 @@ public class MainController {
     }
 
     @QueryMapping(name = "numberOfConvertion")
-    public Integer numberOfConvertion(@Argument String inputCharCode,
+    public Long numberOfConvertion(@Argument String inputCharCode,
                                       @Argument String outputCharCode) {
         return valuteService.getNumberOfConvertion(inputCharCode, outputCharCode);
     }
